isolate email sending logic
