+++ next-email-app/index.js
@@ -1,70 +1,10 @@
 const { app, BrowserWindow, ipcMain } = require("electron");
 const path = require("path");
-const DOMPurify = require("dompurify");
-const { JSDOM } = require("jsdom");
-const validator = require("validator");
-const applescript = require("applescript");
+const sendEmail = require(path.join(__dirname, "sendEmail.js"));
 
-// Setup DOMPurify instance with JSDOM
-const window = new JSDOM("").window;
-const purify = DOMPurify(window);
-
-// Sanitize HTML content
-function sanitizeHTML(input) {
-  return purify.sanitize(input);
-}
-
-// Function to validate email format
-function isValidEmail(email) {
-  return validator.isEmail(email);
-}
-
-function isValidName(name) {
-  return /^[a-zA-Z\s]*$/.test(name);
-}
-
-function containsHTMLTags(input) {
-  const htmlTagRegex = /<[^>]+>/; // Regex to detect HTML tags
-  return htmlTagRegex.test(input); // Returns true if HTML tags are found
-}
-
-// Function to spawn child process for sending email
-function sendEmail(name, email, text) {
-  // Sanitize inputs
-  const safeName = isValidName(name) ? name : "";
-  const safeEmail = isValidEmail(email) ? email : "";
-
-  if (containsHTMLTags(text)) {
-    console.log(`HTML tags are not allowed in text: ${text}.`);
-    return;
-  }
-
-  const formattedText = text.replace(/\n/g, "<br>"); // Replace newlines with breaks.
-  const safeText = sanitizeHTML(formattedText).replace(/"/g, '\\"'); // Sanitize any HTML in the body
-
-  // Spawn a child process for sending the email
-  const script = `
-  set {ccName01, ccAddress01} to {"Example CC", "exampleCC@example.com"} -- 'Cc:' recipient.
-  
-  set the_Subject to "Example Subject"
-
-  set the_Content to ("<div>" & "Hello ${safeName}" & "</div>" & "<br>" & "Thank you for opting in for an email to ${safeEmail}" & "</br>" & "<br>" & "</br>" & "<br>" & "Here is some text:" & "</br>" & "<br>" & "${safeText}" & "</br>" & "<br>" & "</br>" & "<br>" & "</br>" & "<div>" & "Best," & "</div>")
-
-  tell application "Microsoft Outlook"
-      set ComplaintMessage to make new outgoing message with properties {subject:the_Subject, content:the_Content}
-      make new cc recipient at ComplaintMessage with properties {email address:{name:ccName01, address:ccAddress01}}
-      open ComplaintMessage
-  end tell
-`;
-
-  applescript.execString(script, (err, result) => {
-    if (err) {
-      console.error(`Error executing AppleScript: ${err}`);
-      return;
-    }
-    console.log(`AppleScript result: ${result}`);
-  });
-}
+// ~~~~~~~~~~~~~~~~~~
+// APP INITIALIZATION
+// ~~~~~~~~~~~~~~~~~~
 
 function createWindow() {
   const mainWindow = new BrowserWindow({
@@ -76,6 +16,8 @@ function createWindow() {
       enableRemoteModule: false, // Disables remote module
       webviewTag: false, // Block all web view tags
       preload: path.join(__dirname, "preload.js"), // Use a secure preloading script for IPC
+      sandbox: true,
+      disableBlinkFeatures: "AuxClick",
     },
   });
 
@@ -85,36 +27,41 @@ function createWindow() {
   mainWindow.loadURL("file://" + path.join(__dirname, "/out/index.html")); // Adjust based on your app build output
 }
 
-// Listen for the IPC event from the renderer process
+app.commandLine.appendSwitch("disable-network-access"); // Disables network access entirely
+
+app.enableSandbox(); // Enable sandbox for all processes
+
+app.whenReady().then(createWindow); // Electron initialization
+
+// ~~~~~~~~~~~~~~~~~~~
+// APP EVENT LISTENERS
+// ~~~~~~~~~~~~~~~~~~~
+
 ipcMain.on(
   "send-email",
   (event, { sanitizedName, sanitizedEmail, sanitizedText }) => {
     sendEmail(sanitizedName, sanitizedEmail, sanitizedText);
   }
-);
+); // Listen for the IPC event from the renderer process
 
-// Disables network access entirely
-app.commandLine.appendSwitch("disable-network-access");
-
-// Enable sandbox for all processes
-app.enableSandbox();
-
-// Electron initialization
-app.whenReady().then(createWindow);
-
-// Quit when all windows are closed
 app.on("window-all-closed", () => {
   if (process.platform !== "darwin") {
     app.quit();
   }
-});
+}); // Quit when all windows are closed
 
-app.on("activate", () => {
-  if (BrowserWindow.getAllWindows().length === 0) {
-    createWindow();
-  }
-});
+app.on("web-contents-created", (event, contents) => {
+  contents.on("will-navigate", (event, navigationUrl) => {
+    event.preventDefault();
+  });
+}); // Prevent any website page redirects
+
+app.on("web-contents-created", (event, contents) => {
+  contents.setWindowOpenHandler(({ url }) => {
+    return { action: "deny" };
+  });
+}); // Prevent any additional windows from opening
 
 process.on("unhandledRejection", (reason, promise) => {
   console.error("Unhandled Rejection at:", promise, "reason:", reason);
-});
+}); // Logging unhandled rejections
