+++ next-email-app/sendEmail.js
@@ -1,39 +1,67 @@
-const { execFile } = require("child_process");
+const DOMPurify = require("dompurify");
+const { JSDOM } = require("jsdom");
+const validator = require("validator");
+const applescript = require("applescript");
 
-function sendEmailWithOutlook(name, email, text) {
-  const appleScript = `
-    set {ccName01, ccAddress01} to {"Example CC", "exampleCC@example.com"} -- 'Cc:' recipient.
-    
-    set the_Subject to "Example Subject"
+// Setup DOMPurify instance with JSDOM
+const window = new JSDOM("").window;
+const purify = DOMPurify(window);
 
-    set the_Content to ("<div>" & "Hello ${name}" & "</div>" & "<br>" & "Thank you for opting in for an email to ${email}" & "</br>" & "<br>" & "</br>" & "<br>" & "Here is some text:" & "</br>" & "<br>" & "${text}" & "</br>" & "<br>" & "</br>" & "<br>" & "</br>" & "<div>" & "Best," & "</div>")
+// Sanitize HTML content
+function sanitizeHTML(input) {
+  return purify.sanitize(input);
+}
 
-    tell application "Microsoft Outlook"
-        
-    set ComplaintMessage to make new outgoing message with properties {subject:the_Subject, content:the_Content}
-    
-    make new cc recipient at ComplaintMessage with properties {email address:{name:ccName01, address:ccAddress01}}
+// Function to validate email format
+function isValidEmail(email) {
+  return validator.isEmail(email);
+}
 
-    open ComplaintMessage
+function isValidName(name) {
+  return /^[a-zA-Z\s]*$/.test(name);
+}
 
-    end tell
-  `;
+function containsHTMLTags(input) {
+  const htmlTagRegex = /<[^>]+>/; // Regex to detect HTML tags
+  return htmlTagRegex.test(input); // Returns true if HTML tags are found
+}
 
-  execFile("osascript", ["-e", appleScript], (error, stdout, stderr) => {
-    if (error) {
-      console.error(`Error executing AppleScript: ${error.message}`);
+// Function to spawn child process for sending email
+function sendEmail(name, email, text) {
+  // Sanitize inputs
+  const safeName = isValidName(name) ? name : "";
+  const safeEmail = isValidEmail(email) ? email : "";
+
+  if (containsHTMLTags(text)) {
+    console.log(`HTML tags are not allowed in text: ${text}.`);
+    return;
+  }
+
+  const formattedText = text.replace(/\n/g, "<br>"); // Replace newlines with breaks.
+  const safeText = sanitizeHTML(formattedText).replace(/"/g, '\\"'); // Sanitize any HTML in the body
+
+  // Spawn a child process for sending the email
+  const script = `
+  set {ccName01, ccAddress01} to {"Example CC", "exampleCC@example.com"} -- 'Cc:' recipient.
+  
+  set the_Subject to "Example Subject"
+
+  set the_Content to ("<div>" & "Hello ${safeName}" & "</div>" & "<br>" & "Thank you for opting in for an email to ${safeEmail}" & "</br>" & "<br>" & "</br>" & "<br>" & "Here is some text:" & "</br>" & "<br>" & "${safeText}" & "</br>" & "<br>" & "</br>" & "<br>" & "</br>" & "<div>" & "Best," & "</div>")
+
+  tell application "Microsoft Outlook"
+      set ComplaintMessage to make new outgoing message with properties {subject:the_Subject, content:the_Content}
+      make new cc recipient at ComplaintMessage with properties {email address:{name:ccName01, address:ccAddress01}}
+      open ComplaintMessage
+  end tell
+`;
+
+  applescript.execString(script, (err, result) => {
+    if (err) {
+      console.error(`Error executing AppleScript: ${err}`);
       return;
     }
-    if (stderr) {
-      console.error(`AppleScript error: ${stderr}`);
-      return;
-    }
-    console.log(`AppleScript output: ${stdout}`);
+    console.log(`AppleScript result: ${result}`);
   });
 }
 
-// Get arguments from the command line
-const [, , name, email, text] = process.argv;
-
-// Send the email
-sendEmailWithOutlook(name, email, text);
+module.exports = sendEmail;
